console: Refactor offline indicator
