+++ pkg/webui/locales/en.json
@@ -24,6 +24,8 @@
   "components.navigation.side.index.hideSidebar": "Hide sidebar",
   "components.notification.details.index.showDetails": "Show details",
   "components.notification.details.index.hideDetails": "Hide details",
+  "components.offline-status.index.offline": "The application went offline",
+  "components.offline-status.index.online": "The application is back online",
   "components.safe-inspector.index.copied": "Copied to clipboard!",
   "components.safe-inspector.index.toggleVisibility": "Toggle visibility",
   "components.safe-inspector.index.copyClipboard": "Copy to clipboard",
@@ -354,8 +356,6 @@
   "console.views.overview.index.continueWorking": "Walk right through to your applications and/or gateways",
   "console.views.overview.index.componentStatus": "Component status",
   "console.views.overview.index.versionInfo": "Version info",
-  "containers.offline-status.index.offline": "The application went offline",
-  "containers.offline-status.index.online": "The application is back online",
   "lib.components.init.initializing": "Initializing…",
   "lib.components.with-auth.errTooFewRights": "Your account does not possess sufficient rights to use the console.",
   "lib.components.with-auth.errTooFewRightsTitle": "Insufficient rights",
