+++ pkg/webui/locales/xx.json
@@ -24,6 +24,8 @@
   "components.navigation.side.index.hideSidebar": "Xxxx xxxxxxx",
   "components.notification.details.index.showDetails": "Xxxx xxxxxxx",
   "components.notification.details.index.hideDetails": "Xxxx xxxxxxx",
+  "components.offline-status.index.offline": "Xxx xxxxxxxxxxx xxxx xxxxxxx",
+  "components.offline-status.index.online": "Xxx xxxxxxxxxxx xx xxxx xxxxxx",
   "components.safe-inspector.index.copied": "Xxxxxx xx xxxxxxxxx!",
   "components.safe-inspector.index.toggleVisibility": "Xxxxxx xxxxxxxxxx",
   "components.safe-inspector.index.copyClipboard": "Xxxx xx xxxxxxxxx",
@@ -354,8 +356,6 @@
   "console.views.overview.index.continueWorking": "Xxxx xxxxx xxxxxxx xx xxxx xxxxxxxxxxxx xxx/xx xxxxxxxx",
   "console.views.overview.index.componentStatus": "Xxxxxxxxx xxxxxx",
   "console.views.overview.index.versionInfo": "Xxxxxxx xxxx",
-  "containers.offline-status.index.offline": "Xxx xxxxxxxxxxx xxxx xxxxxxx",
-  "containers.offline-status.index.online": "Xxx xxxxxxxxxxx xx xxxx xxxxxx",
   "lib.components.init.initializing": "Xxxxxxxxxxxxx",
   "lib.components.with-auth.errTooFewRights": "Xxxx xxxxxxx xxxx xxx xxxxxxx xxxxxxxxxx xxxxxx xx xxx xxx xxxxxxx.",
   "lib.components.with-auth.errTooFewRightsTitle": "Xxxxxxxxxxxx xxxxxx",
